requirements.txt
tensorflow>=2.9.0
numpy>=1.19.5
Pillow>=8.2.0
scikit-learn>=0.24.2
streamlit>=1.8.0
